+++ ms-notificacion/src/main/resources/db/migration/V1__crear_tabla_notificacion.sql
@@ -1,10 +1,10 @@
 CREATE TABLE notificacion (
-                                id_notificacion INT AUTO_INCREMENT PRIMARY KEY,
-                                id_ticket BIGINT NOT NULL,
-                                correo_destino VARCHAR(100) NOT NULL,
-                                mensaje VARCHAR(255) NOT NULL,
-                                tipo_notificacion VARCHAR(50) NOT NULL, -- Ej: 'PAGO_RECIBIDO', 'EQUIPO_REPARADO'
-                                fecha_envio DATETIME NOT NULL
+                              id_notificacion BIGINT AUTO_INCREMENT PRIMARY KEY,
+                              id_ticket BIGINT NOT NULL,
+                              correo_destino VARCHAR(100) NOT NULL,
+                              mensaje VARCHAR(255) NOT NULL,
+                              tipo_notificacion VARCHAR(50) NOT NULL, -- Ej: 'PAGO_RECIBIDO', 'EQUIPO_REPARADO'
+                              fecha_envio DATETIME NOT NULL
 );
